@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -37,6 +37,7 @@ dependencies:
   provider: ^6.0.5
   dio: ^5.1.2
   http: any
+  web_socket_channel: ^2.4.0
 
   # storage
   shared_preferences: ^2.1.1
@@ -56,13 +57,19 @@ dependencies:
   flutter_switch: ^0.3.2
   syncfusion_flutter_charts: 21.2.4
   permission_handler: ^10.2.0
-  flutter_html: ^2.2.1
   flutter_native_splash: 2.2.0
-  
+  telephony: ^0.2.0
+  geolocator: ^9.0.1
+  url_launcher: ^6.1.5
 
   # other lib
   flutter_localizations:
     sdk: flutter
+  firebase_auth: ^4.16.0
+  equatable: ^2.0.5
+  bloc: ^8.1.3
+  flutter_bloc: ^8.1.4
+  chaquopy: 0.0.20
 
 dependency_overrides:
   # language
