@@ -2,22 +2,13 @@ import 'dart:io';
 
 import 'package:bluetooth_ecg/components/circular_avatar.dart';
 import 'package:bluetooth_ecg/constants/color_constant.dart';
-import 'package:bluetooth_ecg/controllers/news_controller.dart';
-import 'package:bluetooth_ecg/controllers/user_controller.dart';
 import 'package:bluetooth_ecg/providers/auth_provider.dart';
-import 'package:bluetooth_ecg/providers/news_provider.dart';
-import 'package:bluetooth_ecg/screens/bluetooth_screens/ble_screen.dart';
-import 'package:bluetooth_ecg/screens/news_screens/news_all_screens.dart';
-import 'package:bluetooth_ecg/screens/news_screens/news_detail_screen.dart';
-import 'package:bluetooth_ecg/utils/files_management.dart';
-import 'package:bluetooth_ecg/utils/utils.dart';
+import 'package:bluetooth_ecg/screens/preview_calculation.dart';
 import 'package:flutter/material.dart';
-import 'package:bluetooth_ecg/components/live_chart.dart';
 import 'package:flutter_switch/flutter_switch.dart';
-import 'package:permission_handler/permission_handler.dart';
 import 'package:phosphor_flutter/phosphor_flutter.dart';
 import 'package:provider/provider.dart';
-import 'package:intl/intl.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 class HomeScreen extends StatefulWidget {
   const HomeScreen({Key? key}) : super(key: key);
@@ -28,40 +19,58 @@ class HomeScreen extends StatefulWidget {
 
 class _HomeScreenState extends State<HomeScreen> {
   final ScrollController _scrollController = ScrollController();
+
   late File fileToSave;
   bool isShowChart = false;
+
+  final TextEditingController _phoneController = TextEditingController();
+  bool _isEditing = false;
+  final FocusNode focusInputPhone = FocusNode();
+  String phoneNumberWarning = "";
 
   Map allNews = {};
   @override
   void initState() {
     super.initState();
-    getDoctorAssigned();
-    NewsController.getAllNews();
+    _getInitPhoneNumber();
+    // getDoctorAssigned();
+    // NewsController.getAllNews();
   }
 
-  void getDoctorAssigned() async {
-    int patientId = await Utils.getUserId();
-    UserController.getDoctorAssigned(patientId);
-  }
-
-  Future<bool> _requestManageStorage() async {
-    final PermissionStatus status =
-        await Permission.manageExternalStorage.request();
-    if (status == PermissionStatus.granted) {
-      return true;
-    } else {
-      return false;
-    }
-  }
+  // void getDoctorAssigned() async {
+  //   int patientId = await Utils.getUserId();
+  //   UserController.getDoctorAssigned(patientId);
+  // }
 
   @override
   void dispose() {
     super.dispose();
+    _phoneController.dispose();
+    focusInputPhone.dispose();
+  }
+
+  _savePhoneNumberToSharedPrefs(String phoneNumber) async {
+    const String keyToSave = "phone_number_warning";
+    final SharedPreferences preferences = await SharedPreferences.getInstance();
+    preferences.setString(keyToSave, phoneNumber);
+  }
+
+  _getInitPhoneNumber() async {
+    const String keyInit = "phone_number_warning";
+    final SharedPreferences preferences = await SharedPreferences.getInstance();
+
+    final bool isExistPhoneNumber = preferences.containsKey(keyInit);
+    if (isExistPhoneNumber) {
+      _phoneController.text = preferences.getString(keyInit)!;
+    }
   }
 
   @override
   Widget build(BuildContext context) {
-    final List allNews = context.watch<NewsProvider>().allNews;
+    int current = 0;
+    final size = MediaQuery.of(context).size;
+    final height = size.height;
+    final width = size.width;
 
     return Container(
       padding: const EdgeInsets.only(right: 20, left: 20, top: 40, bottom: 10),
@@ -77,37 +86,47 @@ class _HomeScreenState extends State<HomeScreen> {
               child: Row(
                 mainAxisAlignment: MainAxisAlignment.spaceBetween,
                 children: [
-                  Row(children: [
-                    const CircularAvatar(
-                        imageAsset: 'assets/images/doctor.png', radius: 27),
-                    const SizedBox(width: 8),
-                    Column(
-                      mainAxisAlignment: MainAxisAlignment.spaceAround,
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        const Text("Chào mừng đến với fmECG"),
-                        Text(
-                          "Thai Dong",
-                          style: TextStyle(
-                            color: ColorConstant.primary,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        )
-                      ],
+                  // Row(children: [
+                  //   const SizedBox(width: 8),
+                  //   Column(
+                  //     mainAxisAlignment: MainAxisAlignment.spaceAround,
+                  //     crossAxisAlignment: CrossAxisAlignment.start,
+                  //     children: [
+                  //       const Text("Chào mừng đến với fmECG"),
+                  //       Text(
+                  //         "Thai Dong",
+                  //         style: TextStyle(
+                  //           color: ColorConstant.primary,
+                  //           fontWeight: FontWeight.bold,
+                  //         ),
+                  //       )
+                  //     ],
+                  //   ),
+                  // ]),
+                  const IconButton(
+                    icon: Icon(
+                      Icons.menu,
+                      size: 35,
                     ),
-                  ]),
+                    onPressed: null,
+                  ),
                   Row(children: [
-                    const DarkLightSwitch(),
-                    const SizedBox(width: 8),
+                    // const DarkLightSwitch(),
+
                     InkWell(
                       child: Icon(PhosphorIcons.regular.bell),
-                    )
+                    ),
+                    SizedBox(
+                      width: width * 0.05,
+                    ),
+                    const CircularAvatar(
+                        imageAsset: 'assets/images/doctor.png', radius: 27),
                   ])
                 ],
               ),
             ),
 
-            const SizedBox(height: 30),
+            SizedBox(height: height * 0.02),
             SizedBox(
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.center,
@@ -115,185 +134,135 @@ class _HomeScreenState extends State<HomeScreen> {
                   Container(
                     alignment: Alignment.topLeft,
                     child: Text(
-                      "Tổng quan",
+                      "Chào mừng tới fmEcg,\nThai Dong.",
                       style: TextStyle(
-                          color: ColorConstant.quaternary,
+                          color: ColorConstant.black900,
                           fontWeight: FontWeight.bold,
                           fontSize: 22),
                     ),
                   ),
-                  const SizedBox(height: 10),
+                  SizedBox(height: height * 0.05),
                   Wrap(
                     spacing: 15,
                     runSpacing: 15,
                     alignment: WrapAlignment.spaceAround,
                     children: [
                       NumberCard(
-                          number: 82.0,
-                          text: "Nhịp tim (bpm)",
-                          subText: "Mẹ",
-                          color1: ColorConstant.primary,
-                          color2: ColorConstant.primary),
+                        leadingIcon: Icon(
+                          PhosphorIcons.fill.heartbeat,
+                          color: Colors.red.shade400,
+                        ),
+                        centerImage:
+                            Image.asset("assets/images/heart_rate.png"),
+                        number: 82.0,
+                        text: "Nhịp tim",
+                        subText: "Mẹ",
+                        unit: "bpm",
+                        // color1: Colors.grey,
+                        // color2: Colors.grey,
+                      ),
                       NumberCard(
-                          number: 72.9,
-                          text: "Biến thiên nhịp tim (bpm)",
-                          subText: "Mẹ",
-                          color1: ColorConstant.primary,
-                          color2: ColorConstant.quaternary),
+                        leadingIcon: Icon(PhosphorIcons.fill.stethoscope),
+                        centerImage:
+                            Image.asset("assets/images/blood_pressure.png"),
+                        number: 72.9,
+                        text: "Huyết áp",
+                        subText: "Mẹ",
+                        unit: "mmHg",
+                        // color1: Colors.grey,
+                        // color2: Colors.grey,
+                      ),
                       NumberCard(
-                          number: 82.6,
-                          text: "Nhịp tim (bpm)",
-                          subText: "Thai nhi",
-                          color1: ColorConstant.primary,
-                          color2: ColorConstant.primary),
+                        leadingIcon: Icon(PhosphorIcons.fill.heartbeat,
+                            color: Colors.red.shade400),
+                        number: 82.6,
+                        centerImage:
+                            Image.asset("assets/images/heart_rate.png"),
+                        text: "Nhịp tim ",
+                        subText: "Thai nhi",
+                        unit: "bpm",
+                        // color1: ColorConstant.primary,
+                        // color1: Colors.grey,
+                        // color2: Colors.grey,
+                        //  color2: ColorConstant.primary,
+                      ),
                       NumberCard(
-                          number: 86.0,
-                          text: "Biến thiên nhịp tim (bpm)",
-                          subText: "Thai nhi",
-                          color1: ColorConstant.primary,
-                          color2: ColorConstant.quaternary),
+                        centerImage:
+                            Image.asset("assets/images/blood_pressure.png"),
+                        leadingIcon: Icon(PhosphorIcons.fill.stethoscope),
+                        number: 86.0,
+                        text: "Huyết áp",
+                        subText: "Thai nhi",
+                        unit: "mmHg",
+                        //color1: ColorConstant.primary,
+                        // color1: ColorConstant.teal,
+                        // color2: ColorConstant.teal,
+                      ),
                     ],
                   )
                 ],
               ),
             ),
-            const SizedBox(height: 30),
-            SizedBox(
-              child: Column(
-                crossAxisAlignment: CrossAxisAlignment.start,
-                children: [
-                  Text(
-                    "Tín hiệu điện tim",
-                    style: TextStyle(
-                        color: ColorConstant.quaternary,
-                        fontWeight: FontWeight.bold,
-                        fontSize: 22),
-                  ),
-                  const SizedBox(height: 10),
-                  !isShowChart
-                      ? ImageCard(
-                          imageAsset: 'assets/images/heart_rate_example.jpeg',
-                          functionScanBluetooth: () {
-                            Navigator.push(
-                                context,
-                                MaterialPageRoute(
-                                  builder: (context) => const BleReactiveScreen(),
-                                ));
-                          },
-                          temporaryNothing: () async {
-                            bool isAccessFiles = await _requestManageStorage();
-                            if (isAccessFiles) {
-                              FilesManagement
-                                  .createDirectoryFirstTimeWithDevice();
-                              fileToSave = await FilesManagement
-                                  .setUpFileToSaveDataMeasurement();
-                              setState(() {
-                                isShowChart = true;
-                              });
-                            } else {
-                              // show dialog need permission
-                              print('phone does not grant permission');
-                            }
-                          })
-                      : LiveChartSample(fileToSave: fileToSave),
-                ],
+            const SizedBox(height: 20),
+            Container(
+              alignment: Alignment.topLeft,
+              padding: const EdgeInsets.only(bottom: 10),
+              child: Text(
+                "Số người thân",
+                style: TextStyle(
+                  color: ColorConstant.quaternary,
+                  fontWeight: FontWeight.bold,
+                  fontSize: 22,
+                ),
               ),
             ),
 
-            const SizedBox(height: 30),
-            Row(mainAxisAlignment: MainAxisAlignment.spaceBetween, children: [
-              Text(
-                "Các tin tức",
-                style: TextStyle(
-                    color: ColorConstant.quaternary,
-                    fontWeight: FontWeight.bold,
-                    fontSize: 22),
+            // Ô nhập liệu số điện thoại
+            TextFormField(
+              controller: _phoneController,
+              keyboardType: TextInputType.phone,
+              enabled: _isEditing,
+              focusNode: focusInputPhone,
+              decoration: const InputDecoration(
+                border: OutlineInputBorder(
+                    borderSide: BorderSide(color: Colors.black)),
+                hintText: 'Nhập số điện thoại',
+                labelText: 'Số điện thoại',
               ),
-              ElevatedButton(
-                onPressed: () {
-                  Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                          builder: (context) => const NewsAllScreen()));
-                },
-                child: const Text("Xem tất cả"),
-              )
-            ]),
+              onChanged: (text) => setState(() {
+                phoneNumberWarning = _phoneController.text;
+              }),
+            ),
 
-            if (allNews.isNotEmpty)
-              ListView.builder(
-                  padding: const EdgeInsets.only(top: 10),
-                  shrinkWrap: true,
-                  itemCount: 4,
-                  physics: const NeverScrollableScrollPhysics(),
-                  itemBuilder: (context, index) {
-                    final news = allNews[index];
-                    final String imagePresentUrl = news["image"] ?? "";
-                    final int newsId = news["news_id"];
-                    final String newsCategory = news["category_name"];
-                    final DateTime newsCreatedAt =
-                        DateTime.parse(news["created_at"]);
-                    final String newsCreatedAtFormat =
-                        DateFormat("EEEE, dd-MM-yyyy", "vi")
-                            .format(newsCreatedAt);
-                    final String newsTitle = news["title"].length > 100
-                        ? news["title"].substring(0, 100)
-                        : news["title"];
-
-                    return InkWell(
-                      onTap: () async {
-                        await NewsController.getNewsById(newsId);
-                        Navigator.push(
-                            context,
-                            MaterialPageRoute(
-                                builder: (context) =>
-                                    const NewsDetailScreen()));
-                      },
-                      splashColor: ColorConstant.primary,
-                      child: Container(
-                        margin: const EdgeInsets.only(bottom: 10),
-                        child: Row(children: [
-                          ClipRRect(
-                            borderRadius: BorderRadius.circular(10),
-                            child: Image.network(
-                              imagePresentUrl,
-                              width: 90,
-                              height: 90,
-                              fit: BoxFit.cover,
-                            ),
-                          ),
-                          const SizedBox(width: 20),
-                          SizedBox(
-                            height: 90,
-                            // BE CAREFUL: BAD EXPERIENCE WHEN LONG WIDTH
-                            width: 210,
-                            child: Column(
-                                mainAxisAlignment:
-                                    MainAxisAlignment.spaceBetween,
-                                crossAxisAlignment: CrossAxisAlignment.start,
-                                children: [
-                                  Text(newsCategory,
-                                      style: TextStyle(
-                                          fontSize: 14,
-                                          fontWeight: FontWeight.bold,
-                                          color: Colors.grey[600])),
-                                  Text(newsTitle,
-                                      overflow: TextOverflow.ellipsis,
-                                      maxLines: 2,
-                                      style: const TextStyle(
-                                          fontSize: 17,
-                                          fontWeight: FontWeight.bold,
-                                          color: Colors.black)),
-                                  Text(newsCreatedAtFormat,
-                                      style: TextStyle(
-                                        color: Colors.grey[700],
-                                      )),
-                                ]),
-                          )
-                        ]),
-                      ),
-                    );
-                  }),
+            // Nút chỉnh sửa và lưu
+            Row(
+              mainAxisAlignment: MainAxisAlignment.center, // Căn giữa các nút
+              children: [
+                ElevatedButton(
+                  onPressed: () {
+                    setState(() {
+                      _isEditing = true;
+                    });
+                    focusInputPhone.requestFocus();
+                  },
+                  child: const Text('Chỉnh sửa'),
+                ),
+                const SizedBox(width: 20), // Khoảng cách giữa các nút
+                ElevatedButton(
+                  onPressed: phoneNumberWarning == ""
+                      ? null
+                      : () async {
+                          await _savePhoneNumberToSharedPrefs(
+                              _phoneController.text);
+                          setState(() {
+                            _isEditing = false;
+                            phoneNumberWarning = "";
+                          });
+                        },
+                  child: const Text('Lưu'),
+                ),
+              ],
+            ),
           ],
         ),
         // child: LiveChartSample()
@@ -355,9 +324,9 @@ class _DarkLightSwitchState extends State<DarkLightSwitch> {
           isDarkSwitch = val;
         });
         if (isDarkSwitch) {
-          auth.setTheme(ThemeType.DARK, false);
+          auth.setTheme(ThemeType.dark, false);
         } else {
-          auth.setTheme(ThemeType.LIGHT, false);
+          auth.setTheme(ThemeType.light, false);
         }
       },
     );
@@ -368,7 +337,8 @@ class SquareContainer extends StatelessWidget {
   final IconData icon;
   final String text;
 
-  const SquareContainer({Key? key, required this.icon, required this.text}) : super(key: key);
+  const SquareContainer({Key? key, required this.icon, required this.text})
+      : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -402,65 +372,112 @@ class SquareContainer extends StatelessWidget {
   }
 }
 
-class NumberCard extends StatelessWidget {
+class NumberCard extends StatefulWidget {
   final double number;
   final String text;
   final String subText;
-  final Color color1;
-  final Color color2;
+  final String unit;
+  final Icon leadingIcon;
+  final Image centerImage;
+  final Function? onTap;
 
-  const NumberCard({Key? key, 
-    required this.number,
-    required this.text,
-    required this.subText,
-    required this.color1,
-    required this.color2,
-  }) : super(key: key);
+  const NumberCard(
+      {Key? key,
+      required this.number,
+      required this.text,
+      required this.subText,
+      required this.unit,
+      required this.centerImage,
+      required this.leadingIcon,
+      this.onTap})
+      : super(key: key);
+
+  @override
+  State<NumberCard> createState() => _NumberCardState();
+}
+
+class _NumberCardState extends State<NumberCard> {
+  bool _isClicked = false;
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      width: 150.0,
-      height: 100.0,
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(12.0),
-        gradient: LinearGradient(
-          colors: [color1, color2],
-          begin: Alignment.topCenter,
-          end: Alignment.bottomCenter,
+    final size = MediaQuery.of(context).size;
+    final height = size.height;
+    final width = size.width;
+    return GestureDetector(
+      onTap: () {
+        setState(() {
+          _isClicked = !_isClicked;
+        });
+        Navigator.push(
+            context,
+            MaterialPageRoute(
+                builder: (context) => const PreviewCalculation()));
+        widget.onTap?.call();
+      },
+      child: Container(
+        width: width * 0.4,
+        height: height * 0.3,
+        decoration: BoxDecoration(
+          borderRadius: BorderRadius.circular(12.0),
+          gradient: LinearGradient(
+            colors: _isClicked
+                ? [Colors.teal, Colors.tealAccent]
+                : [Colors.grey, Colors.blueGrey],
+            begin: Alignment.topCenter,
+            end: Alignment.bottomCenter,
+          ),
         ),
-      ),
-      child: Column(
-        mainAxisAlignment: MainAxisAlignment.center,
-        crossAxisAlignment: CrossAxisAlignment.center,
-        children: [
-          Text(
-            text,
-            textAlign: TextAlign.center,
-            style: TextStyle(
-              fontWeight: FontWeight.bold,
-              fontSize: 15.0,
-              color: ColorConstant.description,
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.start,
+          children: [
+            ListTile(
+              leading: widget.leadingIcon,
+              title: Text(
+                widget.text,
+                textAlign: TextAlign.center,
+                style: TextStyle(
+                  fontSize: 16,
+                  fontWeight: FontWeight.bold,
+                  color: _isClicked
+                      ? ColorConstant.black900
+                      : ColorConstant.description,
+                ),
+              ),
             ),
-          ),
-          const SizedBox(height: 3.0),
-          Text(
-            subText,
-            style: TextStyle(
-              fontSize: 16.0,
-              color: ColorConstant.description,
+            const SizedBox(height: 3.0),
+            Text(
+              widget.subText,
+              style: TextStyle(
+                fontSize: 16.0,
+                color: _isClicked ? Colors.amber : Colors.brown,
+              ),
             ),
-          ),
-          const SizedBox(height: 5.0),
-          Text(
-            '$number',
-            style: TextStyle(
-              fontSize: 30.0,
-              color: ColorConstant.description,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-        ],
+            const SizedBox(height: 10.0),
+            Expanded(child: widget.centerImage),
+            Row(mainAxisAlignment: MainAxisAlignment.spaceAround, children: [
+              Text(
+                '${widget.number}',
+                style: TextStyle(
+                  fontSize: 30.0,
+                  color: _isClicked
+                      ? ColorConstant.black900
+                      : ColorConstant.description,
+                  fontWeight: FontWeight.bold,
+                ),
+              ),
+              Text(
+                widget.unit,
+                style: TextStyle(
+                  color: _isClicked
+                      ? ColorConstant.black900
+                      : ColorConstant.description,
+                  fontSize: 15,
+                ),
+              ),
+            ]),
+          ],
+        ),
       ),
     );
   }
@@ -471,7 +488,8 @@ class ImageCard extends StatelessWidget {
   final Function() functionScanBluetooth;
   final Function() temporaryNothing;
 
-  const ImageCard({Key? key, 
+  const ImageCard({
+    Key? key,
     required this.imageAsset,
     required this.functionScanBluetooth,
     required this.temporaryNothing,
