@@ -1,8 +1,26 @@
-const express = require('express');
-const authController = require('../controllers/AuthenController');
+const express = require("express");
+const AuthenRouter = require("./AuthenRouter");
+const DeviceRouter = require("./DeviceRouter");
+const UserRouter = require("./UserRouter");
 
-let router = express.Router();
+const HeartRecRouter = require("./HeartRecRouter");
 
-//router.post('/auth/login', authController.login);
+const RecordRouter = require("./RecordRouter");
+const BloodPressureRouter = require("./BloodPressureRouter");
+
+
+const router = express.Router();
+
+router.use("/heart_rec", HeartRecRouter);
+
+router.use("/auth", AuthenRouter);
+
+router.use("/user", UserRouter);
+
+router.use("/device", DeviceRouter);
+
+router.use("/record", RecordRouter);
+
+router.use("/bp", BloodPressureRouter);
 
 module.exports = router;
